+++ CHANGELOG.md
@@ -13,10 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- All reader only provide access to the inner reader by the "into_inner()" method.
-- Changed block size of XZOptions to NonZero type.
+- Renamed LZMA2's "independent work unit" naming from "stream" to "chunk" to not consume it with XZ streams.
 - LZMA2Writer now take a LZMA2Option struct. This enables both the LZMA2Writer and LZMA2WriterMT to encode multiple
-  streams for multi-threaded decoding.
+  chunks for multi-threaded decoding.
+- Changed block size of XZOptions to NonZero type.
+- Unified the API of the writers as far as possible.
 
 ### Fixed